+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 **Packer**
 - Increase minimum supported version to 1.7.
 
+**Terraform**
+- Add `create` parameter to `workspace` method.
+
 ### 2.0.2
 **General**
 - More type checking for required parameters.